refactor: 발주 내역 메서드 DAO 분리
- 클래스별 명확한 역할 분리를 위해 UI의 쿼리 관련 메서드 DAO로 분리
